+++ lib/src/models/anonymous_user_attributes.dart
@@ -7,11 +7,13 @@ class AnonymousUserAttributes {
     this.address,
     this.fields,
     this.languageCode,
+    this.marketId,
     this.timeZone,
   });
   final String? firstName;
   final String? lastName;
   final String? languageCode;
+  final String? marketId;
   final String? timeZone;
   final Address? address;
   final List<UserCustomField>? fields;
@@ -21,6 +23,7 @@ class AnonymousUserAttributes {
       'firstName': firstName,
       'lastName': lastName,
       'languageCode': languageCode,
+      'marketId': marketId,
       'timeZone': timeZone,
       'address': address?.toMap(),
       'fields': fields?.map((e) => e.toMap()).toList(),